+++ CHANGELOG.md
@@ -5,16 +5,27 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.1.0] - 2022-07-30
+
+This release completes the available triggers in the Windows Task Scheduler.
+
 ### Added
 * Add monthly dow trigger
 * Add monthly trigger
 * Add Changelog
+* Add event trigger
+* Add registration trigger
 
 ### Changed
 * Improve documentation
 * Change triggers to accept bools instead of i16
 * Update cargo toml
 * Update readme
+
+### Fixed
+* Fix casting issues
+* Removed mutability to parameters that do not need it
 
 ## [0.0.1] - 2022-04-10
 ### Added